requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.33
fredapi>=0.5.1
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3

